Move the Metadata in its own module
